Changes according to review
