--- examples/display_section.py
@@ -1,9 +1,0 @@
-import xarray as xr
-import xrviz
-from xrviz.display import Display
-from xrviz.sample_data import great_lakes
-
-ds = great_lakes
-
-display = Display(ds)
-display.show()
