+++ xrviz/tests/test_control.py
@@ -1,7 +1,7 @@
 import pytest
 from xrviz.control import Control
 from . import data
-from ..utils import cartopy_geoviews_installed
+from ..compatability import has_cartopy
 
 
 @pytest.fixture()
@@ -11,7 +11,7 @@ def control(data):
 
 def test_control_initial(control):
     tabs = [tab.name for tab in control.tabs]
-    if cartopy_geoviews_installed():
+    if has_cartopy:
         assert tabs == ['Set Coords', 'Variables', 'Axes', 'Style',
                         'Projection']
     else:
